+++ src/components/Navbar.js
@@ -1,6 +1,5 @@
-/* eslint-disable react/prop-types */
-/* eslint-disable react/destructuring-assignment */
 import React from 'react';
+import PropTypes from 'prop-types';
 import { MdMic, MdSettings } from 'react-icons/md';
 import { NavLink } from 'react-router-dom';
 
@@ -27,8 +26,15 @@ const Navbar = (props) => {
         <MdSettings />
       </div>
     </div>
-
   );
+};
+
+Navbar.propTypes = {
+  title: PropTypes.string.isRequired,
+  left: PropTypes.oneOfType([
+    PropTypes.string,
+    PropTypes.element,
+  ]).isRequired,
 };
 
 export default Navbar;
