+++ src/redux/countries/countries.js
@@ -35,11 +35,11 @@ const filterNumbers = (countries, value) => {
     case '99999':
       return countries.filter((country) => country.All.confirmed < 100000);
     case '100000':
-      return countries.filter((country) => country.All.confirmed
-      >= 100000 && country.All.confirmed < 200000);
+      return countries.filter((country) => country.All.confirmed >= 100000
+        && country.All.confirmed < 200000);
     case '200000':
-      return countries.filter((country) => country.All.confirmed
-      >= 200000 && country.All.confirmed < 300000);
+      return countries.filter((country) => country.All.confirmed >= 200000
+        && country.All.confirmed < 300000);
     case '300000':
       return countries.filter((country) => country.All.confirmed >= 300000);
     default:
@@ -65,7 +65,10 @@ const countriesReducer = (state = initialState, action) => {
     case LOAD_CONTINENT:
       return { continent: action.payload };
     case LOAD_COUNTRY:
-      return { ...action.payload };
+      return {
+        ...state,
+        current_country: action.payload,
+      };
     case FILTER_DATA:
       return { continent: action.payload };
     default:
