+++ src/styles/App.css
@@ -1,6 +1,7 @@
 @import url('https://fonts.googleapis.com/css2?family=Lato:wght@300;400;700&display=swap');
 
 :root {
+  --pink: #f03f83;
   --dark-pink: #ec4c8a;
   --light-pink: #fc5193;
 }
@@ -12,7 +13,7 @@ p {
 }
 
 body {
-  background-color: var(--dark-pink);
+  background-color: var(--pink);
 }
 
 .card {
@@ -30,21 +31,12 @@ body {
 }
 
 .filter {
-  background-color: var(--dark-pink);
-}
-
-.map {
-  opacity: 0.3;
-}
-
-.card-body {
-  padding: 0;
-  flex: none;
+  background-color: var(--pink);
 }
 
 .dropdown,
 .dropdown:hover {
-  background-color: var(--dark-pink);
+  background-color: var(--pink);
   border: none;
   outline: none;
 }
@@ -52,8 +44,8 @@ body {
 .show > .btn-primary.dropdown-toggle,
 .show > .btn-primary.dropdown-toggle:focus,
 .btn-primary:focus {
-  background-color: var(--dark-pink);
-  box-shadow: 0 0 0 0 var(--dark-pink);
+  background-color: var(--pink);
+  box-shadow: 0 0 0 0 var(--pink);
   border: none;
 }
 
@@ -69,7 +61,16 @@ a:hover {
 
 .navbar {
   padding: 10px;
-  background-color: var(--dark-pink);
+  background-color: var(--pink);
+}
+
+.map {
+  opacity: 0.3;
+}
+
+.card-body {
+  padding: 0;
+  flex: none;
 }
 
 li:nth-child(odd) {
